// Frontend/src/api/generated/model/actualControlDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { OperationalControlGroupDto } from './operationalControlGroupDto';
import type { OperationalControlFrequencyDto } from './operationalControlFrequencyDto';
import type { OperationalControlStatusDto } from './operationalControlStatusDto';
import type { OperationalControlRevisionDto } from './operationalControlRevisionDto';

export interface ActualControlDto {
  id?: string;
  /** @nullable */
  concept?: string | null;
  /** @nullable */
  control?: string | null;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  lastReview?: string | null;
  /** @nullable */
  nextReview?: string | null;
  group?: OperationalControlGroupDto;
  frequency?: OperationalControlFrequencyDto;
  status?: OperationalControlStatusDto;
  /** @nullable */
  revisions?: OperationalControlRevisionDto[] | null;
}