// Frontend/src/api/generated/model/operationalControlMasterDataDto.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */
import type { OperationalControlYearDto } from './operationalControlYearDto';
import type { OperationalControlGroupDto } from './operationalControlGroupDto';
import type { OperationalControlFrequencyDto } from './operationalControlFrequencyDto';
import type { OperationalControlStatusDto } from './operationalControlStatusDto';

export interface OperationalControlMasterDataDto {
  /** @nullable */
  years?: OperationalControlYearDto[] | null;
  /** @nullable */
  groups?: OperationalControlGroupDto[] | null;
  /** @nullable */
  frequencies?: OperationalControlFrequencyDto[] | null;
  /** @nullable */
  statuses?: OperationalControlStatusDto[] | null;
}