// Frontend/src/api/generated/model/createActualControlRequest.ts
/**
 * Generated by orval v7.8.0 🍺
 * Do not edit manually.
 * TMB Control API
 * OpenAPI spec version: v1
 */

export interface CreateActualControlRequest {
  /** @nullable */
  concept?: string | null;
  /** @nullable */
  control?: string | null;
  /** @nullable */
  observations?: string | null;
  /** @nullable */
  lastReview?: string | null;
  /** @nullable */
  nextReview?: string | null;
  groupId?: number;
  frequencyId?: number;
  statusId?: number;
}